requirements.txt
tensorflow>=2.10
opencv-python
flask
pillow
scikit-learn
matplotlib
